(#23) Changed loop points to a scale of 0.0 to 1.0, added missing docs

+++ src/main/java/tech/fastj/systems/audio/Audio.java
@@ -9,14 +9,30 @@ import java.util.concurrent.TimeUnit;
 /** An audio object used for sound playback. */
 public class Audio {
 
-    /** Signifies that the audio should loop indefinitely when it finishes playing. */
+    /**
+     * Signifies that the audio should loop indefinitely when it finishes playing.
+     * <p>
+     * Made for use with {@link Audio#setLoopCount(int)}.
+     */
     public static final int ContinuousLoop = Clip.LOOP_CONTINUOUSLY;
-    /** Signifies that the audio should stop looping. */
+    /**
+     * Signifies that the audio should stop looping.
+     * <p>
+     * Made for use with {@link Audio#setLoopCount(int)}.
+     */
     public static final int StopLooping = 0;
 
-    /** Signifies that the audio should start its loop from the beginning of the audio track. */
+    /**
+     * Signifies that the audio should start its loop from the beginning of the audio track.
+     * <p>
+     * Made for use with {@link Audio#setLoopPoints(float, float)}, as the first parameter.
+     */
     public static final int LoopFromStart = 0;
-    /** Signifies that the audio should loop back once it reaches the end of the audio track. */
+    /**
+     * Signifies that the audio should loop back once it reaches the end of the audio track.
+     * <p>
+     * Made for use with {@link Audio#setLoopPoints(float, float)}, as the second parameter.
+     */
     public static final int LoopAtEnd = -1;
 
     private final Clip clip;
@@ -24,8 +40,8 @@ public class Audio {
     private final AudioEventListener audioEventListener;
     private final Path audioPath;
 
-    private int loopStart;
-    private int loopEnd;
+    private float loopStart;
+    private float loopEnd;
     private int loopCount;
     boolean shouldLoop;
     PlaybackState currentPlaybackState;
@@ -112,18 +128,44 @@ public class Audio {
         return TimeUnit.MILLISECONDS.convert(clip.getMicrosecondPosition(), TimeUnit.MILLISECONDS);
     }
 
-    public int getLoopStart() {
+    /**
+     * Gets the audio's starting point for looping, on a scale of {@code 0.0} to {@code 1.0}.
+     *
+     * @return The audio's loop start point on a scale of {@code 0.0} to {@code 1.0}.
+     */
+    public float getLoopStart() {
         return loopStart;
     }
 
-    public int getLoopEnd() {
+    /**
+     * Gets the audio's ending point for looping, on a scale of {@code 0.0} to {@code 1.0}.
+     *
+     * @return The audio's loop end point on a scale of {@code 0.0} to {@code 1.0}.
+     */
+    public float getLoopEnd() {
         return loopEnd;
     }
 
+    /**
+     * Gets the audio's loop count -- the amount of times the audio should loop.
+     *
+     * @return The amount of times the audio should loop.
+     */
     public int getLoopCount() {
         return loopCount;
     }
 
+    /**
+     * Gets whether the should should be looping.
+     * <p>
+     * Notes:
+     * <ul>
+     *     <li>This can be affected by {@link Audio#setShouldLoop(boolean)}, {@link Audio#setLoopPoints(float, float)},
+     *     and {@link Audio#setLoopCount(int)} (in most circumstances).</li>
+     * </ul>
+     *
+     * @return Whether the audio should be looping.
+     */
     public boolean shouldLoop() {
         return shouldLoop;
     }
@@ -139,11 +181,20 @@ public class Audio {
 
     /**
      * Sets the audio's loop points.
+     * <p>
+     * Notes:
+     * <ul>
+     *     <li>Both {@code loopStart} and {@code loopEnd} should be on a scale of {@code 0.0} to {@code 1.0}.</li>
+     *     <li>Calling this method automatically tells the audio to loop (you can check this using {@link Audio#shouldLoop()}).
+     *     See {@link Audio#setLoopCount(int)} and {@link Audio#setShouldLoop(boolean)} for ways to disable looping.</li>
+     * </ul>
      *
-     * @param loopStart The point to loop back to; the starting point for the audio loop.
-     * @param loopEnd   The point at which to loop; the ending point for the audio loop.
+     * @param loopStart The point to loop back to; the starting point for the audio loop. This value is on a scale of
+     *                  {@code 0.0} to {@code 1.0}.
+     * @param loopEnd   The point at which to loop; the ending point for the audio loop. This value is on a scale of
+     *                  {@code 0.0} to {@code 1.0}.
      */
-    public void setLoopPoints(int loopStart, int loopEnd) {
+    public void setLoopPoints(float loopStart, float loopEnd) {
         if (loopEnd < -1) {
             throw new IllegalArgumentException("The endpoint for the loop should not be less than -1.");
         }
@@ -162,6 +213,7 @@ public class Audio {
      * Notes:
      * <ul>
      *     <li>If you want to loop continuously, call this method with {@link Audio#ContinuousLoop} as the {@code loopCount} parameter.</li>
+     *     <li>If you want to disable looping, call this method with {@link Audio#StopLooping} as the {@code loopCount} parameter.</li>
      * </ul>
      *
      * @param loopCount The amount of times to loop.
@@ -175,6 +227,18 @@ public class Audio {
         shouldLoop = (this.loopCount != StopLooping);
     }
 
+    /**
+     * Directly sets whether the audio should loop, overriding changes made through {@link Audio#setLoopCount(int)} and
+     * {@link Audio#setLoopPoints(float, float)}.
+     * <p>
+     * Notes:
+     * <ul>
+     *     <li>If the loop count (check using {@link Audio#getLoopCount()}) is set to {@code 0} or {@link Audio#StopLooping},
+     *     calling this method will have no effect on whether the audio will loop.</li>
+     * </ul>
+     *
+     * @param shouldLoop Whether the audio should loop.
+     */
     public void setShouldLoop(boolean shouldLoop) {
         this.shouldLoop = shouldLoop;
     }
@@ -198,6 +262,7 @@ public class Audio {
      * <p>
      * Notes:
      * <ul>
+     *     <li>If the audio was set to loop (check using {@link Audio#shouldLoop()}) then calling this method will cause it to loop.</li>
      *     <li>To stop the audio's playback, use {@link Audio#stop()}.</li>
      *     <li>Starting the audio's playback calls an "audio start" event, which can be hooked into using {@link AudioEventListener#setAudioStartAction(Runnable)}.</li>
      *     <li>Starting the audio's playback calls an "audio open" event, which can be hooked into using {@link AudioEventListener#setAudioOpenAction(Runnable)}.</li>
@@ -224,6 +289,7 @@ public class Audio {
      * <p>
      * Notes:
      * <ul>
+     *     <li>If the audio was set to loop (check using {@link Audio#shouldLoop()}) then calling this method will cause it to continue looping.</li>
      *     <li>Starting the audio's playback calls an "audio resume" event, which can be hooked into using {@link AudioEventListener#setAudioResumeAction(Runnable)}.</li>
      * </ul>
      */
@@ -245,6 +311,7 @@ public class Audio {
         AudioManager.stopAudio(this);
     }
 
+    /** Immediately stops the audio's looping, setting whether the audio should loop immediately to {@code false}. */
     public void stopLoopingNow() {
         clip.loop(StopLooping);
         shouldLoop = false;
